backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uvicorn

chromadb==0.4.24
sentence-transformers==2.5.1
langchain-community==0.0.28
langchain==0.1.12
torch==2.2.1
